+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - webapp: Add folders page to browse media by directory tree
 - webapp: Autoplay videos in the media view. It can be disabled with `webapp.pages.mediaView.autoPlayVideo`
 - webapp: Show the index as root folder on the folders page with `webapp.pages.folders.showIndex`
+- webapp: Add order menu to the nav bar of media lists which sets the `order by` expression of the search query
+- webapp: Add ascending and descending name order to the folders page. Its initial order can be set with `webapp.pages.folders.order`
 
 ### Fixed